+++ crk_orthography/syllabics.py
@@ -81,6 +81,20 @@ def syllabics2sro(syllabics: str, produce_macrons=False) -> str:
     >>> syllabics2sro('Eddie ᓂᑎᓯᔨᐦᑳᓱᐣ᙮')
     'Eddie nitisiyihkâson.'
 
+    You should be able to convert words written in any Western Cree dialect.
+
+    Y-dialect (a.k.a., Plains Cree):
+    >>> syllabics2sro('ᓂᔭ')
+    'niya'
+
+    Th-dialect (a.k.a., Woods Cree):
+    >>> syllabics2sro('ᓂᖬ')
+    'nitha'
+
+    N-dialect (a.k.a., Swampy Cree):
+    >>> syllabics2sro('ᓃᓇ')
+    'nîna'
+
     By default, the SRO will be produced with circumflexes (âêîô):
 
     >>> syllabics2sro('ᐁᐍᐹᐲᐦᑫᐍᐱᓇᒪᕽ')